# openapi.json
{
    "components": {
        "schemas": {
...
            "CreateFamilyTree": {
                "properties": {
                    "display_name": {
                        "description": "Human-readable display name.",
                        "type": "string"
                    },
                    "is_primary": {
                        "description": "Set to `true` to make this the owner's primary tree (clears `is_primary` on all others).",
                        "type": "boolean"
                    },
                    "name": {
                        "description": "Unique slug identifier for this tree (e.g. `\"smith-family\"`). Must be globally unique.",
                        "type": "string"
                    },
                    "owner": {
                        "description": "Username of the tree owner.",
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "display_name",
                    "owner"
                ],
                "type": "object"
            },
...
            "ErrorResponse": {
                "description": "Error response body returned for all 4xx/5xx responses.",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
            "FamilyTree": {
                "properties": {
                    "display_name": {
                        "description": "Human-readable display name.",
                        "type": "string"
                    },
                    "id": {
                        "description": "Record ID in the form `family_tree:<ulid>`.",
                        "example": "family_tree:01jd4a8xyz",
                        "type": "string"
                    },
                    "is_primary": {
                        "description": "Whether this is the owner's primary tree.",
                        "type": "boolean"
                    },
                    "name": {
                        "description": "Unique slug identifier for this tree (e.g. `\"smith-family\"`). Globally unique.",
                        "type": "string"
                    },
                    "owner": {
                        "description": "Username of the tree owner.",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "display_name",
                    "owner"
                ],
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "REST API for ancestry and family tree data management, backed by SurrealDB.",
        "license": {
            "name": ""
        },
        "title": "clann-server API",
        "version": "0.1.0"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/trees": {
            "get": {
                "operationId": "list_trees",
                "parameters": [
                    {
                        "description": "When present, only return trees with this owner.",
                        "in": "query",
                        "name": "owner",
                        "required": false,
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/FamilyTree"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "List of family trees"
                    }
                },
                "tags": [
                    "trees"
                ]
            },
            "post": {
                "operationId": "create_tree",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFamilyTree"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FamilyTree"
                                }
                            }
                        },
                        "description": "Family tree created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Name already taken or bad request"
                    }
                },
                "tags": [
                    "trees"
                ]
            }
        },
        "/api/trees/{name}": {
            "delete": {
                "operationId": "delete_tree",
                "parameters": [
                    {
                        "description": "Unique tree name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Family tree and all its persons deleted"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Family tree not found"
                    }
                },
                "tags": [
                    "trees"
                ]
            },
            "get": {
                "operationId": "get_tree",
                "parameters": [
                    {
                        "description": "Unique tree name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FamilyTree"
                                }
                            }
                        },
                        "description": "Family tree found"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Family tree not found"
                    }
                },
                "tags": [
                    "trees"
                ]
            }
        }
    },
...
}